test(components): mock LAPIS calls for tests #44
