+++ components/index.html
@@ -8,7 +8,7 @@
         <script src="./node_modules/lit/polyfill-support.js"></script>
         <script type="module" src="./src/components/app.ts"></script>
         <script type="module" src="./src/components/location-filter.ts"></script>
-        <script type="module" src="./src/components/prevalence-over-time.ts"></script>
+        <script type="module" src="src/components/prevalenceOverTime/prevalence-over-time.ts"></script>
     </head>
     <body>
         <gs-app lapis="https://lapis.cov-spectrum.org/open/v1/sample">
