+++ components/package.json
@@ -16,7 +16,8 @@
         "storybook": "storybook dev -p 6006",
         "build-storybook": "storybook build",
         "test:storybook": "test-storybook",
-        "test:playwright": "playwright test"
+        "test:playwright": "playwright test",
+        "test:playwright:update-snapshots": "playwright test --update-snapshots"
     },
     "keywords": [
         "web-components",
@@ -47,7 +48,6 @@
         "@typescript-eslint/eslint-plugin": "^7.1.1",
         "@typescript-eslint/parser": "^7.1.1",
         "autoprefixer": "^10.4.18",
-        "concurrently": "^8.2.2",
         "daisyui": "^4.7.3",
         "eslint": "^8.57.0",
         "eslint-plugin-storybook": "^0.8.0",
@@ -56,6 +56,7 @@
         "postcss": "^8.4.35",
         "prettier": "^3.2.5",
         "storybook": "^7.6.17",
+        "storybook-addon-fetch-mock": "^1.0.1",
         "tailwindcss": "^3.4.1",
         "typescript": "~5.4.2",
         "vitest": "^1.3.1"
