+++ components/src/components/prevalenceOverTime/prevalence-over-time-bubble-chart.ts
@@ -1,12 +1,12 @@
 import { customElement, property } from 'lit/decorators.js';
 import { html, LitElement } from 'lit';
 import { Chart, registerables } from 'chart.js';
-import { TemporalGranularity } from '../types';
-import { addUnit, minusTemporal } from '../temporal';
-import { getMinMaxNumber } from '../utils';
-import { getYAxisScale, ScaleType } from './container/component-scaling-selector';
-import { LogitScale } from './charts/LogitScale';
-import { PrevalenceOverTimeData } from '../query/queryPrevalenceOverTime';
+import { TemporalGranularity } from '../../types';
+import { addUnit, minusTemporal } from '../../temporal';
+import { getMinMaxNumber } from '../../utils';
+import { getYAxisScale, ScaleType } from '../container/component-scaling-selector';
+import { LogitScale } from '../charts/LogitScale';
+import { PrevalenceOverTimeData } from '../../query/queryPrevalenceOverTime';
 
 @customElement('gs-prevalence-over-time-bubble-chart')
 export class PrevalenceOverTimeBubbleChart extends LitElement {
