+++ components/src/components/relativeGrowthAdvantage/__mockData__/denominator.json
@@ -1,0 +1,376 @@
+{
+    "errors": [],
+    "info": {
+        "apiVersion": 1,
+        "dataVersion": 1709685650,
+        "deprecationDate": null,
+        "deprecationInfo": null,
+        "acknowledgement": null
+    },
+    "data": [
+        {
+            "date": "2021-01-30",
+            "count": 179
+        },
+        {
+            "date": "2021-01-31",
+            "count": 132
+        },
+        {
+            "date": "2021-02-01",
+            "count": 314
+        },
+        {
+            "date": "2021-02-02",
+            "count": 210
+        },
+        {
+            "date": "2021-02-03",
+            "count": 168
+        },
+        {
+            "date": "2021-02-04",
+            "count": 168
+        },
+        {
+            "date": "2021-02-05",
+            "count": 236
+        },
+        {
+            "date": "2021-02-06",
+            "count": 133
+        },
+        {
+            "date": "2021-02-07",
+            "count": 111
+        },
+        {
+            "date": "2021-02-08",
+            "count": 136
+        },
+        {
+            "date": "2021-02-09",
+            "count": 133
+        },
+        {
+            "date": "2021-02-10",
+            "count": 87
+        },
+        {
+            "date": "2021-02-11",
+            "count": 133
+        },
+        {
+            "date": "2021-02-12",
+            "count": 115
+        },
+        {
+            "date": "2021-02-13",
+            "count": 136
+        },
+        {
+            "date": "2021-02-14",
+            "count": 88
+        },
+        {
+            "date": "2021-02-15",
+            "count": 121
+        },
+        {
+            "date": "2021-02-16",
+            "count": 133
+        },
+        {
+            "date": "2021-02-17",
+            "count": 157
+        },
+        {
+            "date": "2021-02-18",
+            "count": 75
+        },
+        {
+            "date": "2021-02-19",
+            "count": 114
+        },
+        {
+            "date": "2021-02-20",
+            "count": 68
+        },
+        {
+            "date": "2021-02-21",
+            "count": 97
+        },
+        {
+            "date": "2021-02-22",
+            "count": 154
+        },
+        {
+            "date": "2021-02-23",
+            "count": 153
+        },
+        {
+            "date": "2021-02-24",
+            "count": 118
+        },
+        {
+            "date": "2021-02-25",
+            "count": 93
+        },
+        {
+            "date": "2021-02-26",
+            "count": 53
+        },
+        {
+            "date": "2021-02-27",
+            "count": 61
+        },
+        {
+            "date": "2021-02-28",
+            "count": 48
+        },
+        {
+            "date": "2021-03-01",
+            "count": 145
+        },
+        {
+            "date": "2020-12-01",
+            "count": 84
+        },
+        {
+            "date": "2020-12-02",
+            "count": 172
+        },
+        {
+            "date": "2020-12-03",
+            "count": 134
+        },
+        {
+            "date": "2020-12-04",
+            "count": 48
+        },
+        {
+            "date": "2020-12-05",
+            "count": 25
+        },
+        {
+            "date": "2020-12-06",
+            "count": 35
+        },
+        {
+            "date": "2020-12-07",
+            "count": 250
+        },
+        {
+            "date": "2020-12-08",
+            "count": 101
+        },
+        {
+            "date": "2020-12-09",
+            "count": 128
+        },
+        {
+            "date": "2020-12-10",
+            "count": 25
+        },
+        {
+            "date": "2020-12-11",
+            "count": 28
+        },
+        {
+            "date": "2020-12-12",
+            "count": 116
+        },
+        {
+            "date": "2020-12-13",
+            "count": 110
+        },
+        {
+            "date": "2020-12-14",
+            "count": 252
+        },
+        {
+            "date": "2020-12-15",
+            "count": 189
+        },
+        {
+            "date": "2020-12-16",
+            "count": 299
+        },
+        {
+            "date": "2020-12-17",
+            "count": 281
+        },
+        {
+            "date": "2020-12-18",
+            "count": 228
+        },
+        {
+            "date": "2020-12-19",
+            "count": 243
+        },
+        {
+            "date": "2020-12-20",
+            "count": 128
+        },
+        {
+            "date": "2020-12-21",
+            "count": 148
+        },
+        {
+            "date": "2020-12-22",
+            "count": 144
+        },
+        {
+            "date": "2020-12-23",
+            "count": 289
+        },
+        {
+            "date": "2020-12-24",
+            "count": 245
+        },
+        {
+            "date": "2020-12-25",
+            "count": 180
+        },
+        {
+            "date": "2020-12-26",
+            "count": 213
+        },
+        {
+            "date": "2020-12-27",
+            "count": 212
+        },
+        {
+            "date": "2020-12-28",
+            "count": 423
+        },
+        {
+            "date": "2020-12-29",
+            "count": 290
+        },
+        {
+            "date": "2020-12-30",
+            "count": 292
+        },
+        {
+            "date": "2020-12-31",
+            "count": 174
+        },
+        {
+            "date": "2021-01-01",
+            "count": 243
+        },
+        {
+            "date": "2021-01-02",
+            "count": 200
+        },
+        {
+            "date": "2021-01-03",
+            "count": 200
+        },
+        {
+            "date": "2021-01-04",
+            "count": 360
+        },
+        {
+            "date": "2021-01-05",
+            "count": 283
+        },
+        {
+            "date": "2021-01-06",
+            "count": 146
+        },
+        {
+            "date": "2021-01-07",
+            "count": 190
+        },
+        {
+            "date": "2021-01-08",
+            "count": 287
+        },
+        {
+            "date": "2021-01-09",
+            "count": 228
+        },
+        {
+            "date": "2021-01-10",
+            "count": 218
+        },
+        {
+            "date": "2021-01-11",
+            "count": 360
+        },
+        {
+            "date": "2021-01-12",
+            "count": 236
+        },
+        {
+            "date": "2021-01-13",
+            "count": 248
+        },
+        {
+            "date": "2021-01-14",
+            "count": 216
+        },
+        {
+            "date": "2021-01-15",
+            "count": 267
+        },
+        {
+            "date": "2021-01-16",
+            "count": 199
+        },
+        {
+            "date": "2021-01-17",
+            "count": 131
+        },
+        {
+            "date": "2021-01-18",
+            "count": 392
+        },
+        {
+            "date": "2021-01-19",
+            "count": 257
+        },
+        {
+            "date": "2021-01-20",
+            "count": 302
+        },
+        {
+            "date": "2021-01-21",
+            "count": 258
+        },
+        {
+            "date": "2021-01-22",
+            "count": 255
+        },
+        {
+            "date": "2021-01-23",
+            "count": 150
+        },
+        {
+            "date": "2021-01-24",
+            "count": 149
+        },
+        {
+            "date": "2021-01-25",
+            "count": 315
+        },
+        {
+            "date": "2021-01-26",
+            "count": 250
+        },
+        {
+            "date": "2021-01-27",
+            "count": 278
+        },
+        {
+            "date": "2021-01-28",
+            "count": 166
+        },
+        {
+            "date": "2021-01-29",
+            "count": 222
+        }
+    ]
+}
