+++ components/src/components/relativeGrowthAdvantage/__mockData__/numerator.json
@@ -1,0 +1,332 @@
+{
+    "errors": [],
+    "info": {
+        "apiVersion": 1,
+        "dataVersion": 1709685650,
+        "deprecationDate": null,
+        "deprecationInfo": null,
+        "acknowledgement": null
+    },
+    "data": [
+        {
+            "date": "2021-01-30",
+            "count": 72
+        },
+        {
+            "date": "2021-01-31",
+            "count": 53
+        },
+        {
+            "date": "2021-02-01",
+            "count": 153
+        },
+        {
+            "date": "2021-02-02",
+            "count": 82
+        },
+        {
+            "date": "2021-02-03",
+            "count": 80
+        },
+        {
+            "date": "2021-02-04",
+            "count": 75
+        },
+        {
+            "date": "2021-02-05",
+            "count": 119
+        },
+        {
+            "date": "2021-02-06",
+            "count": 66
+        },
+        {
+            "date": "2021-02-07",
+            "count": 51
+        },
+        {
+            "date": "2021-02-08",
+            "count": 73
+        },
+        {
+            "date": "2021-02-09",
+            "count": 66
+        },
+        {
+            "date": "2021-02-10",
+            "count": 55
+        },
+        {
+            "date": "2021-02-11",
+            "count": 75
+        },
+        {
+            "date": "2021-02-12",
+            "count": 69
+        },
+        {
+            "date": "2021-02-13",
+            "count": 75
+        },
+        {
+            "date": "2021-02-14",
+            "count": 47
+        },
+        {
+            "date": "2021-02-15",
+            "count": 67
+        },
+        {
+            "date": "2021-02-16",
+            "count": 66
+        },
+        {
+            "date": "2021-02-17",
+            "count": 82
+        },
+        {
+            "date": "2021-02-18",
+            "count": 29
+        },
+        {
+            "date": "2021-02-19",
+            "count": 64
+        },
+        {
+            "date": "2021-02-20",
+            "count": 46
+        },
+        {
+            "date": "2021-02-21",
+            "count": 79
+        },
+        {
+            "date": "2021-02-22",
+            "count": 108
+        },
+        {
+            "date": "2021-02-23",
+            "count": 109
+        },
+        {
+            "date": "2021-02-24",
+            "count": 80
+        },
+        {
+            "date": "2021-02-25",
+            "count": 60
+        },
+        {
+            "date": "2021-02-26",
+            "count": 34
+        },
+        {
+            "date": "2021-02-27",
+            "count": 51
+        },
+        {
+            "date": "2021-02-28",
+            "count": 44
+        },
+        {
+            "date": "2021-03-01",
+            "count": 112
+        },
+        {
+            "date": "2020-12-07",
+            "count": 2
+        },
+        {
+            "date": "2020-12-11",
+            "count": 1
+        },
+        {
+            "date": "2020-12-12",
+            "count": 2
+        },
+        {
+            "date": "2020-12-13",
+            "count": 1
+        },
+        {
+            "date": "2020-12-14",
+            "count": 1
+        },
+        {
+            "date": "2020-12-16",
+            "count": 5
+        },
+        {
+            "date": "2020-12-17",
+            "count": 1
+        },
+        {
+            "date": "2020-12-18",
+            "count": 1
+        },
+        {
+            "date": "2020-12-19",
+            "count": 1
+        },
+        {
+            "date": "2020-12-21",
+            "count": 6
+        },
+        {
+            "date": "2020-12-22",
+            "count": 3
+        },
+        {
+            "date": "2020-12-23",
+            "count": 1
+        },
+        {
+            "date": "2020-12-24",
+            "count": 5
+        },
+        {
+            "date": "2020-12-25",
+            "count": 2
+        },
+        {
+            "date": "2020-12-26",
+            "count": 10
+        },
+        {
+            "date": "2020-12-27",
+            "count": 3
+        },
+        {
+            "date": "2020-12-28",
+            "count": 9
+        },
+        {
+            "date": "2020-12-29",
+            "count": 9
+        },
+        {
+            "date": "2020-12-30",
+            "count": 15
+        },
+        {
+            "date": "2020-12-31",
+            "count": 9
+        },
+        {
+            "date": "2021-01-01",
+            "count": 39
+        },
+        {
+            "date": "2021-01-02",
+            "count": 8
+        },
+        {
+            "date": "2021-01-03",
+            "count": 14
+        },
+        {
+            "date": "2021-01-04",
+            "count": 21
+        },
+        {
+            "date": "2021-01-05",
+            "count": 22
+        },
+        {
+            "date": "2021-01-06",
+            "count": 13
+        },
+        {
+            "date": "2021-01-07",
+            "count": 15
+        },
+        {
+            "date": "2021-01-08",
+            "count": 22
+        },
+        {
+            "date": "2021-01-09",
+            "count": 38
+        },
+        {
+            "date": "2021-01-10",
+            "count": 50
+        },
+        {
+            "date": "2021-01-11",
+            "count": 34
+        },
+        {
+            "date": "2021-01-12",
+            "count": 18
+        },
+        {
+            "date": "2021-01-13",
+            "count": 16
+        },
+        {
+            "date": "2021-01-14",
+            "count": 15
+        },
+        {
+            "date": "2021-01-15",
+            "count": 30
+        },
+        {
+            "date": "2021-01-16",
+            "count": 17
+        },
+        {
+            "date": "2021-01-17",
+            "count": 18
+        },
+        {
+            "date": "2021-01-18",
+            "count": 36
+        },
+        {
+            "date": "2021-01-19",
+            "count": 44
+        },
+        {
+            "date": "2021-01-20",
+            "count": 42
+        },
+        {
+            "date": "2021-01-21",
+            "count": 30
+        },
+        {
+            "date": "2021-01-22",
+            "count": 28
+        },
+        {
+            "date": "2021-01-23",
+            "count": 31
+        },
+        {
+            "date": "2021-01-24",
+            "count": 39
+        },
+        {
+            "date": "2021-01-25",
+            "count": 59
+        },
+        {
+            "date": "2021-01-26",
+            "count": 72
+        },
+        {
+            "date": "2021-01-27",
+            "count": 85
+        },
+        {
+            "date": "2021-01-28",
+            "count": 58
+        },
+        {
+            "date": "2021-01-29",
+            "count": 93
+        }
+    ]
+}
