+++ components/src/constants.ts
@@ -1,1 +1,5 @@
 export const LAPIS_URL = 'https://lapis.cov-spectrum.org/open/v1/sample';
+
+export const AGGREGATED_ENDPOINT = `${LAPIS_URL}/aggregated`;
+export const NUCLEOTIDE_MUTATIONS_ENDPOINT = `${LAPIS_URL}/nuc-mutations`;
+export const NUCLEOTIDE_INSERTIONS_ENDPOINT = `${LAPIS_URL}/nuc-insertions`;
